requirements.txt
python-twitter
boto3
python-dotenv
pynamodb